정답 버튼 1번 눌렀을 시 반응
